files: Move functions that interact with files to their own module.
In working to port over the existing tools to the new common options /
context stuff the omnibus common module makes this particularly
difficult. Breaking up the common module is generally a positive change
regardless.
