+++ src/common.h
@@ -31,6 +31,8 @@
 #ifndef COMMON_H
 #define COMMON_H
 
+#include "files.h"
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -51,13 +53,7 @@ int getSizeUint32Hex(const char *arg, UINT32 *num);
 int getPcrId(const char *arg, UINT32 *num);
 int str2ByteStructure(const char *inStr, UINT16 *byteLenth, BYTE *byteBuffer);
 int hex2ByteStructure(const char *inStr, UINT16 *byteLenth, BYTE *byteBuffer);
-int saveDataToFile(const char *fileName, UINT8 *buf, UINT16 size);
-int loadDataFromFile(const char *fileName, UINT8 *buf, UINT16 *size);
-int saveTpmContextToFile(TSS2_SYS_CONTEXT *sysContext, TPM_HANDLE handle, const char *fileName);
-int loadTpmContextFromFile(TSS2_SYS_CONTEXT *sysContext, TPM_HANDLE *handle, const char *fileName);
 int computeDataHash(BYTE *buffer, UINT16 length, TPMI_ALG_HASH halg, TPM2B_DIGEST *result);
-int checkOutFile(const char *path);
-int getFileSize(const char *path, long *fileSize);
 int getPort(const char *arg, int *port);
 int getDebugLevel(const char *arg, int *dl);
 int parsePCRSelections(const char *arg, TPML_PCR_SELECTION *pcrSels);
